Fix BlueB /merge issue
For some reason the BlueB image is unexpectedly closed prior to merging
causing an error. Simply creating a sacrificial blueB image saves the
original blueB image and allows the script to run. The issue is likely
due to all the close(); commands, but this is a quick fix that allows
the script to run